term-to-fhir-format/src/pyproject.toml
[tool.poetry]
name = "TMT-TMLT-FHIR-Concept-transformer"
version = "0.1.0"
description = ""
authors = ["Pawat <pawat@hlabconsulting.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
pandas = "^1.4.1"
numpy = "^1.22.2"
xlrd = "^2.0.1"
google-cloud-storage = "^2.6.0"
ndjson = "^0.3.1"
google-auth = "^2.14.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.1.0"
bandit = "^1.7.2"
safety = "^1.10.3"
pre-commit = "^2.17.0"
jupyter = "^1.0.0"
jupyterthemes = "^0.20.0"
mypy = "^0.940"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
